added mood transition an dminor fixes

+++ backend/accounts/models.py
@@ -11,7 +11,7 @@ class MoodVector(models.Model):
     playlist_url = models.CharField(max_length=200, null=True)
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
 
-    x_start = models.IntegerField()
-    y_start = models.IntegerField()
-    x_end = models.IntegerField()
-    y_end = models.IntegerField
+    x_start = models.FloatField()
+    y_start = models.FloatField()
+    x_end = models.FloatField()
+    y_end = models.FloatField()
